+++ TKV.Model/Migrations/MyDbContextModelSnapshot.cs
@@ -66,8 +66,8 @@ namespace TKV.Model.Migrations
 
                     SqlServerPropertyBuilderExtensions.UseIdentityColumn(b.Property<int>("Id"));
 
-                    b.Property<int>("Budget")
-                        .HasColumnType("int");
+                    b.Property<double>("Budget")
+                        .HasColumnType("float");
 
                     b.Property<int>("CoverageId")
                         .HasColumnType("int");
